+++ src/permissions/object_permissions.py
@@ -1,0 +1,13 @@
+from rest_framework.permissions import (
+    BasePermission,
+    SAFE_METHODS
+)
+
+
+class IsOwnerOrReadOnly(BasePermission):
+
+    def has_object_permission(self, request, view, obj):
+        if request.method in SAFE_METHODS:
+            return True
+        else:
+            return request.user.id == obj.publisher.id
